requirements.txt
numpy==1.20.2
Pillow==8.4.0
